+++ src/main/resources/config/config-template.conf
@@ -393,7 +393,7 @@ simona.event.listener = [
 # Congestion Management Configuration
 ##################################################################
 
-simona.congestionManagement.enable = boolean | false
+simona.congestionManagement.enableDetection = boolean | false
 simona.congestionManagement.enableTransformerTapping = boolean | false
 simona.congestionManagement.enableTopologyChanges = boolean | false
 simona.congestionManagement.enableUsingFlexOptions = boolean | false
